moves openNewTab into a utils file and uses it for Desktop
